+++ src/Adapters.Tests/Shared/AdapterHelper.cs
@@ -18,7 +18,7 @@ namespace Microsoft.ApplicationInsights.Tracing.Tests
     {
         public readonly string InstrumentationKey;
 
-#if NET40 || NET45 || NET46
+#if NET45 || NET46
         private static readonly string ApplicationInsightsConfigFilePath =
             Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "ApplicationInsights.config");
 #else