+++ src/Adapters/TraceListener.Shared/AssemblyInfo.cs
@@ -31,7 +31,6 @@ using System.Runtime.InteropServices;
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("3aada9bf-fc15-42fd-ace5-4cb2f11c1f2a")]
 
-[assembly: InternalsVisibleTo("Microsoft.ApplicationInsights.TraceListener.Net40.Tests, PublicKey=" + AssemblyInfo.PublicKey)]
 [assembly: InternalsVisibleTo("Microsoft.ApplicationInsights.TraceListener.Net45.Tests, PublicKey=" + AssemblyInfo.PublicKey)]
 
 internal static class AssemblyInfo